+++ tests/Functional/DoctrineOrm/DoctrineOrmRulerTest.php
@@ -128,6 +128,29 @@ class DoctrineOrmRulerTest extends TestCase
         ], $parameters);
     }
 
+    #[Test]
+    public function shouldNotDuplicateJoinOnRepeatedApply(): void
+    {
+        $qb = (new QueryBuilder($this->entityManager))
+            ->from(Product::class, 'products')
+            ->select('products');
+
+        $this->ruler->apply($qb, 'category.key = :key', ['key' => 'foo']);
+        $this->ruler->apply($qb, 'category.enabled = :enabled', ['enabled' => true]);
+
+        /** @var array<string, array<Join>> $joins */
+        $joins = $qb->getDQLPart('join');
+
+        self::assertCount(1, $joins['products']);
+        self::assertEquals('products.category', $joins['products'][0]->getJoin());
+        self::assertEquals('category', $joins['products'][0]->getAlias());
+
+        // The generated DQL must be valid - a duplicated join alias would throw here.
+        $qb->getQuery();
+
+        $this->addToAssertionCount(1);
+    }
+
     public static function provideDataForApply(): array
     {
         return [